dia_01/cu-01/rules-decision-tree/requirements.txt
certifi==2026.2.25
charset-normalizer==3.4.7
ibmcloudant==0.11.5
ibm-cloud-sdk-core==3.24.4
idna==3.11
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
urllib3==2.6.3
